Implement crowd query
